Safari-609.1.11.1/LayoutTests/inspector/debugger/tail-deleted-frames/tail-deleted-frames-intermediate-tail-deleted-frames-expected.txt

Ensure proper tail deleted frames with different call stacks.


== Running test suite: Debugger.TailDeletedFrames.IntermediateTailDeletedFrames
-- Running test case: Debugger.TailDeletedFrames.IntermediateTailDeletedFrames
PAUSED
CALL STACK:
0: [F] a
1: -F- b
2: -F- c
3: [F] startABC
4: [P] Global Code
RESUMED
